MOSTLY FIXES taoneill/war#569
I realize that this is not a perfect implementation of a completed
version of this, but this is about 5 hours of work. I realize that this
is not tested and there may be typos but I don't want to look at the
word "BlockFace" for a while. It is impossible to get everything in one
pass.